unify/app/requirements.txt
blinker==1.9.0
chardet==5.2.0
click==8.1.8
distro==1.9.0
Flask==3.1.0
Flask-Cors==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jpype1==1.5.1
lxml==5.3.0
MarkupSafe==3.0.2
numpy==2.2.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-pptx==1.0.2
pytz==2024.2
reportlab==4.2.5
six==1.17.0
tabula-py==2.10.0
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.1.3
XlsxWriter==3.2.0
